+++ pkg/local_object_storage/metabase/ec_test.go
@@ -148,12 +148,12 @@ func TestDB_ResolveECPart(t *testing.T) {
 			require.NoError(t, db.Put(&tomb))
 		}},
 		{name: "garbage mark only", assertErr: assertObjectNotFoundError, preset: func(t *testing.T, db *meta.DB) {
-			_, _, err := db.MarkGarbage(partAddr)
+			_, err := db.MarkGarbage(partAddr)
 			require.NoError(t, err)
 		}},
 		{name: "stored with garbage mark", assertErr: assertObjectNotFoundError, preset: func(t *testing.T, db *meta.DB) {
 			require.NoError(t, db.Put(&partObj))
-			_, _, err := db.MarkGarbage(parentAddr)
+			_, err := db.MarkGarbage(parentAddr)
 			require.NoError(t, err)
 		}},
 		{name: "container garbage mark only", assertErr: assertObjectNotFoundError, preset: func(t *testing.T, db *meta.DB) {
@@ -171,7 +171,7 @@ func TestDB_ResolveECPart(t *testing.T) {
 		}},
 		{name: "expired with garbage mark", assertErr: assertObjectExpiredError, preset: func(t *testing.T, db *meta.DB) {
 			require.NoError(t, db.Put(&expiredObj))
-			_, _, err := db.MarkGarbage(partAddr)
+			_, err := db.MarkGarbage(partAddr)
 			require.NoError(t, err)
 		}},
 		{name: "expired with container garbage mark", assertErr: assertObjectNotFoundError, preset: func(t *testing.T, db *meta.DB) {
@@ -263,7 +263,7 @@ func TestDB_ResolveECPart(t *testing.T) {
 		{name: "stored with garbage mark and locker", preset: func(t *testing.T) *meta.DB {
 			db := newDB(t)
 			require.NoError(t, db.Put(&partObj))
-			_, _, err := db.MarkGarbage(partAddr)
+			_, err := db.MarkGarbage(partAddr)
 			require.NoError(t, err)
 			require.NoError(t, db.Put(&locker))
 			return db
@@ -785,7 +785,7 @@ func testMarkGarbageEC(t *testing.T) {
 
 	assertECGroupAvailable(t, db, parent, parts)
 
-	inhumed, _, err := db.MarkGarbage(parentAddr)
+	inhumed, err := db.MarkGarbage(parentAddr)
 	require.NoError(t, err)
 
 	allAddrs := append(partAddrs, parentAddr)
@@ -811,7 +811,7 @@ func testMarkGarbageEC(t *testing.T) {
 	require.NoError(t, err)
 	require.ElementsMatch(t, g, append(partAddrs, parentAddr))
 
-	require.EqualValues(t, len(parts), inhumed) // parent is virtual so it doesn't count
+	require.EqualValues(t, len(parts)+1, inhumed) // parent also has a GC mark in the shard now
 }
 
 func assertECGroupAvailable(t *testing.T, db *meta.DB, parent object.Object, parts []object.Object) {