+++ pkg/local_object_storage/shard/inhume.go
@@ -27,10 +27,7 @@ func (s *Shard) MarkGarbage(addrs ...oid.Address) error {
 		return ErrDegradedMode
 	}
 
-	var err error
-
-	_, _, err = s.metaBase.MarkGarbage(addrs...)
-
+	inhumed, err := s.metaBase.MarkGarbage(addrs...)
 	if err != nil {
 		s.log.Debug("could not mark object to delete in metabase",
 			zap.Error(err),
@@ -38,6 +35,8 @@ func (s *Shard) MarkGarbage(addrs ...oid.Address) error {
 
 		return fmt.Errorf("metabase inhume: %w", err)
 	}
+
+	s.addObjectCounter(gcObjType, int(inhumed))
 
 	if s.hasWriteCache() {
 		for i := range addrs {